// src/api/user/types.ts
/* eslint-disable */
// tslint:disable
/**
 * Yeying API
 * 
 *
 * OpenAPI spec version: 1.0
 * Contact: yeying.community@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator+.
 * https://github.com/karlvr/openapi-generator-plus
 * Do not edit the class manually.
 */

import { Express } from 'express'
import { Api } from '../../models'

export interface UserApi {
	userAdd: (request: Api.UserAddUserRequest) => Promise<UserAddResponse>
	userDelete: (request: Api.UserDeleteUserRequest) => Promise<UserDeleteResponse>
	userDetail: (request: Api.UserUserDetailRequest) => Promise<UserDetailResponse>
	userList: (request: Api.UserUserListRequest) => Promise<UserListResponse>
	userUpdate: (request: Api.UserUpdateUserRequest) => Promise<UserUpdateResponse>
	userUpdateStatus: (request: Api.UserUpdateStatusRequest) => Promise<UserUpdateStatusResponse>
}

export type UserAddResponse = UserAdd200Response | UserAddDefaultResponse

export interface UserAdd200Response {
	status: 200
	body: Api.UserAddUserResponse
	headers?: never
}

export interface UserAddDefaultResponse {
	/* Catch-all response */
	status: 'default'
	actualStatus: number
	body: Api.RpcStatus
	headers?: never
}

export type UserDeleteResponse = UserDelete200Response | UserDeleteDefaultResponse

export interface UserDelete200Response {
	status: 200
	body: Api.UserDeleteUserResponse
	headers?: never
}

export interface UserDeleteDefaultResponse {
	/* Catch-all response */
	status: 'default'
	actualStatus: number
	body: Api.RpcStatus
	headers?: never
}

export type UserDetailResponse = UserDetail200Response | UserDetailDefaultResponse

export interface UserDetail200Response {
	status: 200
	body: Api.UserUserDetailResponse
	headers?: never
}

export interface UserDetailDefaultResponse {
	/* Catch-all response */
	status: 'default'
	actualStatus: number
	body: Api.RpcStatus
	headers?: never
}

export type UserListResponse = UserList200Response | UserListDefaultResponse

export interface UserList200Response {
	status: 200
	body: Api.UserUserListResponse
	headers?: never
}

export interface UserListDefaultResponse {
	/* Catch-all response */
	status: 'default'
	actualStatus: number
	body: Api.RpcStatus
	headers?: never
}

export type UserUpdateResponse = UserUpdate200Response | UserUpdateDefaultResponse

export interface UserUpdate200Response {
	status: 200
	body: Api.UserUpdateUserResponse
	headers?: never
}

export interface UserUpdateDefaultResponse {
	/* Catch-all response */
	status: 'default'
	actualStatus: number
	body: Api.RpcStatus
	headers?: never
}

export type UserUpdateStatusResponse = UserUpdateStatus200Response | UserUpdateStatusDefaultResponse

export interface UserUpdateStatus200Response {
	status: 200
	body: Api.UserUpdateStatusResponse
	headers?: never
}

export interface UserUpdateStatusDefaultResponse {
	/* Catch-all response */
	status: 'default'
	actualStatus: number
	body: Api.RpcStatus
	headers?: never
}