// src/yeying/api/user/user.ts
export enum UserStatusEnum {
  /** USER_STATUS_UNKNOWN - 未知编码，是一种占位符 */
  USER_STATUS_UNKNOWN = 0,
  /** USER_STATUS_ACTIVE - 活跃：表示用户目前正在使用软件并与其进行交互 */
  USER_STATUS_ACTIVE = 1,
  /** USER_STATUS_OFFLINE - 离线：表示用户当前不处于活跃状态，但可能会随时返回 */
  USER_STATUS_OFFLINE = 2,
  /** USER_STATUS_DISABLE - 禁用：表示用户的访问权限被暂时或永久性地限制 */
  USER_STATUS_DISABLE = 3,
  /** USER_STATUS_LOCK - 锁定：表示用户的账户被暂时性地冻结，通常是由于安全原因 */
  USER_STATUS_LOCK = 4,
  /** USER_STATUS_UNVERIFIED - 未验证：表示用户账户在创建后尚未通过验证 */
  USER_STATUS_UNVERIFIED = 5,
  /** USER_STATUS_DELETED - 删除：表示用户账户已经被永久删除，无法恢复 */
  USER_STATUS_DELETED = 6,
  /** USER_STATUS_DORMANT - 休眠：表示用户账户由于长时间未活跃而被标记为休眠状态 */
  USER_STATUS_DORMANT = 7,
  /** USER_STATUS_FREEZE - 冻结：表示用户账户在一段时间内暂时不可用 */
  USER_STATUS_FREEZE = 8,
  /** USER_STATUS_AUDIT - 待审核：表示新用户注册后正在等待管理员或审核人员审批 */
  USER_STATUS_AUDIT = 9,
  /** USER_STATUS_REFUSED - 审核未通过：表示管理员或审核人员拒绝了用户的注册请求 */
  USER_STATUS_REFUSED = 10,
  UNRECOGNIZED = -1,
}
// ...
export enum UserRoleEnum {
  /** USER_ROLE_UNKNOWN - 未知编码，是一种占位符 */
  USER_ROLE_UNKNOWN = 0,
  /** USER_ROLE_OWNER - 所有者 */
  USER_ROLE_OWNER = 1,
  /** USER_ROLE_NORMAL - 普通用户 */
  USER_ROLE_NORMAL = 2,
  UNRECOGNIZED = -1,
}